feat: Enhance user deletion and registration logic; add refresh token management

+++ jwt_and_refresh_token_authentication/src/main/java/com/bootcamp/jwt_and_refresh_token_authentication/application/handler/command/user/deletebyemail/DeleteUserByEmailCommandHandler.java
@@ -15,18 +15,21 @@ public class DeleteUserByEmailCommandHandler {
     }
 
     public Result<Void> handle(DeleteUserByEmailCommand command) {
-        try {
-            User user = _userService.getByEmail(command.email());
-            
-            // Yetki kontrolü: Sadece admin silebilir ya da kullanıcı kendi hesabını silebilir.
-            if (!command.isAdmin() && !command.email().equals(command.currentSessionEmail())) {
-                return Result.failure("You don't have permission to delete this user");
-            }
+        User user = _userService.getByEmail(command.email());
+        if (user == null) {
+            return Result.failure("User not found with email: " + command.email());
+        }
 
+        // Yetki kontrolü: Sadece admin silebilir ya da kullanıcı kendi hesabını silebilir.
+        if (!command.isAdmin() && !command.email().equals(command.currentSessionEmail())) {
+            return Result.failure("You don't have permission to delete this user");
+        }
+
+        try {
             _userService.delete(user.getId());
             return Result.success();
         } catch (Exception e) {
-            return Result.failure("User not found with email: " + command.email());
+            return Result.failure("An error occurred while deleting the user: " + e.getMessage());
         }
     }
 }
